COMP: Need full definitions not forward declarations
