+++ Libs/Widgets/ctkPathListWidget.cpp
@@ -31,52 +31,6 @@
 // QtGUI includes
 #include "ctkPathListWidget.h"
 
-// --------------------------------------------------------------------------
-// ctkPathListWidgetPrivate
-
-//-----------------------------------------------------------------------------
-class ctkPathListWidgetPrivate
-{
-  Q_DECLARE_PUBLIC(ctkPathListWidget)
-  Q_GADGET
-
-protected:
-  ctkPathListWidget* const q_ptr;
-
-public:
-
-  enum PathType {
-    Unknown,
-    File,
-    Directory
-  };
-
-  ctkPathListWidgetPrivate(ctkPathListWidget& object);
-
-  void _q_emitPathClicked(const QModelIndex &index);
-  void _q_emitPathDoubleClicked(const QModelIndex &index);
-  void _q_emitPathActivated(const QModelIndex &index);
-  void _q_emitCurrentPathChanged(const QModelIndex &current, const QModelIndex &previous);
-
-  bool addPath(const QString& path);
-  bool removePath(const QString& path);
-
-  void fileOptionsChanged();
-  void directoryOptionsChanged();
-
-  PathType pathType(const QString& absolutePath) const;
-
-  bool isValidPath(const QString& absoluteFilePath, PathType pathType) const;
-  bool isValidFile(const QString& absoluteFilePath) const;
-  bool isValidDir(const QString& absoluteDirPath) const;
-
-  QStandardItemModel PathListModel;
-  ctkPathListWidget::Mode Mode;
-  ctkPathListWidget::PathOptions FileOptions;
-  ctkPathListWidget::PathOptions DirectoryOptions;
-  QIcon FileIcon;
-  QIcon DirectoryIcon;
-};
 
 // --------------------------------------------------------------------------
 // ctkPathListWidgetPrivate methods
