+++ Libs/Widgets/ctkPathListWidget.h
@@ -26,13 +26,10 @@
 
 // Qt includes
 #include <QListView>
+#include <QStandardItem>
 
 // QtGUI includes
 #include "ctkWidgetsExport.h"
-
-class ctkPathListWidgetPrivate;
-
-class QStandardItem;
 
 /// \ingroup Widgets
 ///
@@ -44,6 +41,7 @@ class QStandardItem;
 /// by setting file and directory options.
 /// \sa ctkPathLineEdit, ctkDirectoryButton
 ///
+class ctkPathListWidgetPrivate; // Forward declaration within this file
 class CTK_WIDGETS_EXPORT ctkPathListWidget : public QListView
 {
   Q_OBJECT
@@ -315,6 +313,53 @@ private:
   Q_PRIVATE_SLOT(d_func(), void _q_emitPathDoubleClicked(const QModelIndex& index))
   Q_PRIVATE_SLOT(d_func(), void _q_emitPathActivated(const QModelIndex& index))
   Q_PRIVATE_SLOT(d_func(), void _q_emitCurrentPathChanged(const QModelIndex &previous, const QModelIndex &current))
+};
+
+
+// --------------------------------------------------------------------------
+// ctkPathListWidgetPrivate
+
+//-----------------------------------------------------------------------------
+class ctkPathListWidgetPrivate
+{
+  Q_DECLARE_PUBLIC(ctkPathListWidget)
+
+protected:
+  ctkPathListWidget* const q_ptr;
+
+public:
+
+  enum PathType {
+    Unknown,
+    File,
+    Directory
+  };
+
+  ctkPathListWidgetPrivate(ctkPathListWidget& object);
+
+  void _q_emitPathClicked(const QModelIndex &index);
+  void _q_emitPathDoubleClicked(const QModelIndex &index);
+  void _q_emitPathActivated(const QModelIndex &index);
+  void _q_emitCurrentPathChanged(const QModelIndex &current, const QModelIndex &previous);
+
+  bool addPath(const QString& path);
+  bool removePath(const QString& path);
+
+  void fileOptionsChanged();
+  void directoryOptionsChanged();
+
+  PathType pathType(const QString& absolutePath) const;
+
+  bool isValidPath(const QString& absoluteFilePath, PathType pathType) const;
+  bool isValidFile(const QString& absoluteFilePath) const;
+  bool isValidDir(const QString& absoluteDirPath) const;
+
+  QStandardItemModel PathListModel;
+  ctkPathListWidget::Mode Mode;
+  ctkPathListWidget::PathOptions FileOptions;
+  ctkPathListWidget::PathOptions DirectoryOptions;
+  QIcon FileIcon;
+  QIcon DirectoryIcon;
 };
 
 Q_DECLARE_OPERATORS_FOR_FLAGS(ctkPathListWidget::PathOptions)
